requirements.txt
patool
cv
tensorflow
matplotlib
pillow
pickle
numpy